+++ src/System/RequirementsManager.php
@@ -35,7 +35,6 @@
 
 namespace Glpi\System;
 
-use Glpi\System\Requirement\DataDirectoriesProtectedPath;
 use Glpi\System\Requirement\DbEngine;
 use Glpi\System\Requirement\DbTimezones;
 use Glpi\System\Requirement\DirectoriesWriteAccess;
@@ -49,7 +48,6 @@ use Glpi\System\Requirement\LogsWriteAccess;
 use Glpi\System\Requirement\MemoryLimit;
 use Glpi\System\Requirement\MysqliMysqlnd;
 use Glpi\System\Requirement\PhpVersion;
-use Glpi\System\Requirement\SafeDocumentRoot;
 use Glpi\System\Requirement\SeLinux;
 use Glpi\System\Requirement\SessionsConfiguration;
 use Glpi\System\Requirement\SessionsSecurityConfiguration;
@@ -138,11 +136,6 @@ class RequirementsManager
 
         // Below requirements are optionals
 
-        $safe_doc_root_requirement = new SafeDocumentRoot();
-        $requirements[] = $safe_doc_root_requirement;
-        if (!$safe_doc_root_requirement->isValidated()) {
-            $requirements[] = new DataDirectoriesProtectedPath(Variables::getDataDirectoriesConstants());
-        }
         $requirements[] = new SessionsSecurityConfiguration();
         $requirements[] = new Extension(
             'exif',